+++ spec/suspenders_spec.rb
@@ -6,6 +6,7 @@ RSpec.describe Suspenders do
   end
 
   describe Suspenders::CLI do
+    let(:template_path) { File.expand_path("../lib/templates/web.rb", __dir__) }
     let(:stub_success) {
       allow_any_instance_of(Suspenders::CLI).to receive(:system).and_return(true)
     }
@@ -15,9 +16,9 @@ RSpec.describe Suspenders do
     end
 
     describe ".run" do
-      it "calls rails new with expected arguments" do
-        template_path = File.expand_path("../lib/templates/web.rb", __dir__)
+      it "calls rails new with Heroku arguments by default" do
         args = [
+          {"SUSPENDERS_PAAS" => "heroku"},
           "rails",
           "new",
           "app_name",
@@ -32,10 +33,39 @@ RSpec.describe Suspenders do
         Suspenders::CLI.run("app_name")
       end
 
+      it "calls rails new with Railway arguments when paas is railway" do
+        args = [
+          {"SUSPENDERS_PAAS" => "railway"},
+          "rails",
+          "new",
+          "app_name",
+          "-d=postgresql",
+          "--skip-test",
+          "--skip-kamal",
+          "--skip-docker",
+          "--skip-thruster",
+          "-m=#{template_path}"
+        ]
+
+        expect_any_instance_of(Suspenders::CLI).to receive(:system).with(*args)
+
+        Suspenders::CLI.run("app_name", paas: "railway")
+      end
+
       it "returns true" do
         result = Suspenders::CLI.run("app_name")
 
         expect(result).to eq true
+      end
+
+      context "when paas is unknown" do
+        it "raises without calling rails" do
+          expect_any_instance_of(Suspenders::CLI).not_to receive(:system)
+
+          expect {
+            Suspenders::CLI.run("app_name", paas: "fly")
+          }.to raise_error(Suspenders::Error, 'Unknown PaaS "fly". Expected one of: heroku, railway')
+        end
       end
 
       context "when rails does not exist" do
@@ -58,9 +88,9 @@ RSpec.describe Suspenders do
 
       context "when rails fails to install" do
         let(:app_name) { "app_name" }
-        let(:template_path) { File.expand_path("../lib/templates/web.rb", __dir__) }
         let(:args) {
           [
+            {"SUSPENDERS_PAAS" => "heroku"},
             "rails",
             "new",
             app_name,